Stop copying `0harmony.dll` into the release
This marks the nuget harmony library as private, for build only, so that
`0harmony.dll` is no longer copied into the build target.  Since the mod
uses the shipped-from-workshop harmony mod, this is how things should
be.

I also bumped the harmony version to 2.2.1 to match the current steam
harmony mod.  Not strictly necessary, but since 2.2.1 brings back
`CodeMatcher`, definitely nice to have for that vastly smoother
Transpiler patch writing experience.